refactor: domain 로직을 분류할 domain 디렉터리 생성
