+++ src/view/Printer.js
@@ -1,6 +1,9 @@
-import OutputView from './view/OutputView.js';
-import { RESULT_MARK, WINNER_DELIMITER } from './constants/PrinterConstants.js';
-import { OutputFormat, OutputMessage } from './constants/OutputMessage.js';
+import OutputView from './OutputView.js';
+import {
+  RESULT_MARK,
+  WINNER_DELIMITER,
+} from '../constants/PrinterConstants.js';
+import { OutputFormat, OutputMessage } from '../constants/OutputMessage.js';
 
 class Printer {
   static printHeader(message) {
